+++ src/phoenix/server/api/input_types/add_examples_to_dataset_input.py
@@ -5,7 +5,7 @@ from strawberry import UNSET
 from strawberry.relay import GlobalID
 from strawberry.scalars import JSON
 
-from .DatasetExampleInput import DatasetExampleInput
+from .dataset_example_input import DatasetExampleInput
 
 
 @strawberry.input
