+++ src/phoenix/server/api/input_types/delete_annotations_input.py
