+++ src/phoenix/server/api/types/EmbeddingDimension.py
@@ -31,12 +31,11 @@ from phoenix.pointcloud.clustering import Hdbscan
 from phoenix.pointcloud.pointcloud import PointCloud
 from phoenix.pointcloud.projectors import Umap
 from phoenix.server.api.context import Context
-from phoenix.server.api.input_types import TimeRange
+from phoenix.server.api.input_types import Granularity, TimeRange
 from phoenix.server.api.types.Cluster import to_gql_clusters
 from phoenix.server.api.types.InferencesRole import AncillaryInferencesRole, InferencesRole
 from phoenix.server.api.types.VectorDriftMetricEnum import VectorDriftMetric
 
-from ..input_types.Granularity import Granularity
 from .DataQualityMetric import DataQualityMetric
 from .EmbeddingMetadata import EmbeddingMetadata
 from .Event import create_event_id, unpack_event_id
